pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jekyll-academic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-frontmatter>=1.1.0",
    "requests>=2.33.1",
    "PyYAML>=6.0",
]

[project.scripts]
fetch-external = "scripts.fetch_external:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["scripts*"]
